기능 : 로그인, 마이페이지, 강의, 리뷰 Commit 내용 : 전체적인 리팩토링 진행 및 CRUD 테스트 완료 (강의 상세 페이지 조회 시 Table Not Found Error 발생)

+++ Coding_Galaxy_BE/src/main/java/com/sparta/coding_galaxy_be/service/ReviewService.java
@@ -2,7 +2,6 @@ package com.sparta.coding_galaxy_be.service;
 
 import com.sparta.coding_galaxy_be.dto.requestDto.ReviewRequestDto;
 import com.sparta.coding_galaxy_be.entity.Courses;
-import com.sparta.coding_galaxy_be.entity.KakaoMemberDetailsImpl;
 import com.sparta.coding_galaxy_be.entity.KakaoMembers;
 import com.sparta.coding_galaxy_be.entity.Reviews;
 import com.sparta.coding_galaxy_be.repository.CourseRepository;
@@ -21,46 +20,56 @@ public class ReviewService {
     private final ReviewRepository reviewRepository;
     private final CourseRepository courseRepository;
 
-    public ResponseEntity<?> createReview(Long courseId, ReviewRequestDto reviewRequestDto, KakaoMembers kakaoMembers) {
+    public ResponseEntity<?> createReview(Long courseId, ReviewRequestDto reviewRequestDto, KakaoMembers kakaoMember) {
 
         Courses course = courseRepository.findById(courseId).orElseThrow(
                 () -> new RuntimeException("강의가 존재하지 않습니다.")
         );
 
-        Reviews review = Reviews.builder().
-                star(reviewRequestDto.getStar()).
-                comment(reviewRequestDto.getComment()).
-                kakaoMember(kakaoMembers).
-                course(course).
-                build();
+        Reviews review = Reviews.builder()
+                .star(reviewRequestDto.getStar())
+                .comment(reviewRequestDto.getComment())
+                .kakaoMember(kakaoMember)
+                .course(course)
+                .build();
 
         reviewRepository.save(review);
 
-        return new ResponseEntity<>(HttpStatus.OK);
+        return new ResponseEntity<>("리뷰가 등록되었습니다.", HttpStatus.OK);
     }
 
     @Transactional
-    public ResponseEntity<?> editReview(Long reviewId, ReviewRequestDto reviewRequestDto, KakaoMembers kakaoMembers) {
+    public ResponseEntity<?> editReview(Long courseId, Long reviewId, ReviewRequestDto reviewRequestDto, KakaoMembers kakaoMember) {
 
-        Reviews updateReview = reviewRepository.findById(reviewId).orElseThrow(()
-            -> new NullPointerException("리뷰가 없습니다."));
+        if (!courseRepository.existsByCourseId(courseId)) throw new RuntimeException("강의가 존재하지 않습니다.");
 
-        // 작성자와 로그인 사용자 비교 로직
-
-        updateReview.updateReview(
-                reviewRequestDto.getStar(),
-                reviewRequestDto.getComment()
+        Reviews review = reviewRepository.findById(reviewId).orElseThrow(
+                () -> new RuntimeException("리뷰가 존재하지 않습니다.")
         );
 
-        return new ResponseEntity<>(HttpStatus.OK);
+        if(!kakaoMember.getName().equals(review.getKakaoMember().getName())){
+            throw new RuntimeException("작성자가 아닙니다.");
+        }
+
+        review.updateReview(reviewRequestDto);
+
+        return new ResponseEntity<>("리뷰 수정이 완료되었습니다.", HttpStatus.OK);
     }
 
-    public ResponseEntity<?> removeReview(Long reviewId, KakaoMembers kakaoMembers) {
+    public ResponseEntity<?> deleteReview(Long courseId, Long reviewId, KakaoMembers kakaoMember) {
 
-        // 작성자와 로그인 사용자 비교 로직
+        if (!courseRepository.existsByCourseId(courseId)) throw new RuntimeException("강의가 존재하지 않습니다.");
+
+        Reviews review = reviewRepository.findById(reviewId).orElseThrow(
+                () -> new RuntimeException("리뷰가 존재하지 않습니다.")
+        );
+
+        if(!kakaoMember.getName().equals(review.getKakaoMember().getName())){
+            throw new RuntimeException("작성자가 아닙니다.");
+        }
 
         reviewRepository.deleteById(reviewId);
 
-        return new ResponseEntity<>(HttpStatus.OK);
+        return new ResponseEntity<>("리뷰 삭제가 완료되었습니다.", HttpStatus.OK);
     }
 }
